Automated Prayer/requirements.txt
web3==5.31.3
python-dotenv==1.0.0
sendgrid==6.9.7